Add documentation

+++ src/main/java/sn/api/requests/IsFriendsRequest.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class IsFriendsRequest {
 
     @JsonProperty("user_ids")
-    List<Integer> userIds;
+    List<Long> userIds;
 }
